requirements.txt
dynaconf==3.1.7
numpy==1.21.4
paho-mqtt==1.6.1
Pillow==8.4.0
pkg-resources==0.0.0
sense-hat==2.2.0
